refactor : User, Search ControllerTest 분리
