+++ src/main/java/com/woowacourse/zzinbros/user/web/controller/rest/SearchController.java
@@ -1,18 +1,18 @@
-package com.woowacourse.zzinbros.user.web.controller;
+package com.woowacourse.zzinbros.user.web.controller.rest;
 
 import com.woowacourse.zzinbros.user.dto.UserResponseDto;
 import com.woowacourse.zzinbros.user.service.SearchService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Set;
 
-@Controller
+@RestController
 @RequestMapping("/users")
 public class SearchController {
     private final SearchService searchService;
